Split checkbox components out of components/common.js
