+++ src/components/common.js
@@ -14,7 +14,6 @@ import Link from 'src/components/common/Link'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { centeredSpinner, icon } from 'src/components/icons'
 import { TextInput } from 'src/components/input'
-import Interactive from 'src/components/Interactive'
 import Modal from 'src/components/Modal'
 import { MiniSortable } from 'src/components/table'
 import TopBar from 'src/components/TopBar'
@@ -33,42 +32,10 @@ import * as Utils from 'src/libs/utils'
 
 
 export * from 'src/components/common/buttons'
+export * from 'src/components/common/Checkbox'
 export * from 'src/components/common/IdContainer'
 export { Clickable, Link }
 
-
-export const Checkbox = ({ checked, onChange, disabled, ...props }) => {
-  useLabelAssert('Checkbox', { ...props, allowId: true })
-  return h(Interactive, _.merge({
-    as: 'span',
-    className: 'fa-layers fa-fw',
-    role: 'checkbox',
-    'aria-checked': checked,
-    onClick: () => !disabled && onChange?.(!checked),
-    style: { verticalAlign: 'middle' },
-    disabled
-  }, props), [
-    icon('squareSolid', { style: { color: Utils.cond([disabled, () => colors.light(1.2)], [checked, () => colors.accent()], () => 'white') } }), // bg
-    !disabled && icon('squareLight', { style: { color: checked ? colors.accent(1.2) : colors.dark(0.75) } }), // border
-    checked && icon('check', { size: 8, style: { color: disabled ? colors.dark(0.75) : 'white' } }) // check
-  ])
-}
-
-export const LabeledCheckbox = ({ checked, onChange, disabled, children, ...props }) => {
-  return h(IdContainer, [id => h(Fragment, [
-    h(Checkbox, { checked, onChange, disabled, 'aria-labelledby': id, ...props }),
-    span({
-      id,
-      style: {
-        verticalAlign: 'middle',
-        color: disabled ? colors.dark(0.8) : undefined,
-        cursor: disabled ? 'default' : 'pointer'
-      },
-      onClick: () => !disabled && onChange?.(!checked),
-      disabled
-    }, [children])
-  ])])
-}
 
 export const RadioButton = ({ text, name, labelStyle, ...props }) => {
   return h(IdContainer, [id => h(Fragment, [
